[JB][OPS-11836] fix edge case where user enters a p800 ref greater than Int.Max (#225)

+++ app/edh/RequestModel.scala
@@ -78,7 +78,7 @@ object ServiceID {
 
 final case class PaymentData(
     paymentAmount: Option[BigDecimal],
-    paymentNumber: Option[Int]
+    paymentNumber: Option[Long]
 )
 
 object PaymentData {
